Implement object based config with default object-assign moduleSpecifier
This commit is based on the new API proposed by @jaydenseric in his PR newoga/babel-plugin-transform-replace-object-assign#3. His PR primarily adds support for babel 7. The PR also introduces some necessary breaking changes to the plugin configuration. This commit takes his new API and backports it to the babel 6 supported version of this plugin to ease migrations. The API changes are:

1. Configuration is now optional. If no config is provided, `object-assign` is used.
2. String based configuration is no longer supported. It must be an object with a `moduleSpecifier` key.

Documentation was updated and a new test fixture was addded to test the new default configuration behavior.